+++ Movie/Controllers/FavoritesController.cs
@@ -8,9 +8,11 @@ using System.Web;
 using System.Web.Mvc;
 using Movie.Models;
 using PagedList;
+using Movie.App_Start;
 
 namespace Movie.Controllers
 {
+    [CheckLogin]
     public class FavoritesController : Controller
     {
         private MovieContext db = new MovieContext();
@@ -18,53 +20,76 @@ namespace Movie.Controllers
         // GET: Favorites
         public ActionResult Index(string UserId, string VideoId,int ?page)
         {
-            var favorite = from m in db.Favorites select m;
-            if (!String.IsNullOrEmpty(UserId))
+            if (Cookies.CheckPrivilege() == true)
             {
-                int uid = int.Parse(UserId);
-                favorite = favorite.Where(s => s.UserId == uid);
+                var favorite = from m in db.Favorites select m;
+                if (!String.IsNullOrEmpty(UserId))
+                {
+                    int uid = int.Parse(UserId);
+                    favorite = favorite.Where(s => s.UserId == uid);
+                }
+                if (!String.IsNullOrEmpty(VideoId))
+                {
+                    int vid = int.Parse(VideoId);
+                    favorite = favorite.Where(s => s.VideoId == vid);
+                }
+                List<Total> total = new List<Total>();
+                foreach (var item in favorite)
+                {
+                    Total detail = new Total();
+                    detail.favorite = item;
+                    var user = db.Users.Where(c => c.UserId == item.UserId).FirstOrDefault();
+                    detail.user = user;
+                    var video = db.Videos.Where(c => c.VideoId == item.VideoId).FirstOrDefault();
+                    detail.video = video;
+                    total.Add(detail);
+                }
+                int pagenumber = page ?? 1;
+                int pagesize = 10;
+                IPagedList<Total> pagelist = total.ToPagedList(pagenumber, pagesize);
+                return View(pagelist);
             }
-            if (!String.IsNullOrEmpty(VideoId))
+            else
             {
-                int vid = int.Parse(VideoId);
-                favorite = favorite.Where(s => s.VideoId == vid);
+                return RedirectToAction("Index", "UserOperation");
             }
-            List<Total> total = new List<Total>();
-            foreach (var item in favorite)
-            {
-                Total detail = new Total();
-                detail.favorite = item;
-                var user = db.Users.Where(c => c.UserId == item.UserId).FirstOrDefault();
-                detail.user = user;
-                var video = db.Videos.Where(c => c.VideoId == item.VideoId).FirstOrDefault();
-                detail.video = video;
-                total.Add(detail);
-            }
-            int pagenumber = page ?? 1;
-            int pagesize = 10;
-            IPagedList<Total> pagelist = total.ToPagedList(pagenumber, pagesize);
-            return View(pagelist);
+            
         }
 
         // GET: Favorites/Details/5
         public ActionResult Details(int? id)
         {
-            if (id == null)
+            if (Cookies.CheckPrivilege() == true)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                Favorite favorite = db.Favorites.Find(id);
+                if (favorite == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(favorite);
             }
-            Favorite favorite = db.Favorites.Find(id);
-            if (favorite == null)
+            else
             {
-                return HttpNotFound();
+                return RedirectToAction("Index", "UserOperation");
             }
-            return View(favorite);
+            
         }
 
         // GET: Favorites/Create
         public ActionResult Create()
         {
-            return View();
+            if (Cookies.CheckPrivilege() == true)
+            {
+                return View();
+            }
+            else
+            {
+                return RedirectToAction("Index", "UserOperation");
+            }
         }
 
         // POST: Favorites/Create
@@ -72,33 +97,49 @@ namespace Movie.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "FavoriteId,UserId,VideoId,FavoriteTime")] Favorite favorite)
         {
-            if (ModelState.IsValid)
+            if (Cookies.CheckPrivilege() == true)
             {
-                // 如果是，则在评论表取得收藏表中最大的历史记录编号
-                var MaxId = db.Favorites.Any() ? db.Favorites.Max(p => p.FavoriteId) : 0;
-                // 将取得最大评论编号加一赋值给将要创建的收藏
-                favorite.FavoriteId = MaxId + 1;
-                db.Favorites.Add(favorite);
-                db.SaveChanges();
-                return RedirectToAction("Index");
-            }
+                if (ModelState.IsValid)
+                {
+                    // 如果是，则在评论表取得收藏表中最大的历史记录编号
+                    var MaxId = db.Favorites.Any() ? db.Favorites.Max(p => p.FavoriteId) : 0;
+                    // 将取得最大评论编号加一赋值给将要创建的收藏
+                    favorite.FavoriteId = MaxId + 1;
+                    db.Favorites.Add(favorite);
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
 
-            return View(favorite);
+                return View(favorite);
+            }
+            else
+            {
+                return RedirectToAction("Index", "UserOperation");
+            }
+            
         }
 
         // GET: Favorites/Edit/5
         public ActionResult Edit(int? id)
         {
-            if (id == null)
+            if (Cookies.CheckPrivilege() == true)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                Favorite favorite = db.Favorites.Find(id);
+                if (favorite == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(favorite);
             }
-            Favorite favorite = db.Favorites.Find(id);
-            if (favorite == null)
+            else
             {
-                return HttpNotFound();
+                return RedirectToAction("Index", "UserOperation");
             }
-            return View(favorite);
+            
         }
 
         // POST: Favorites/Edit/5
@@ -106,28 +147,44 @@ namespace Movie.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "FavoriteId,UserId,VideoId,FavoriteTime")] Favorite favorite)
         {
-            if (ModelState.IsValid)
+            if (Cookies.CheckPrivilege() == true)
             {
-                db.Entry(favorite).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                if (ModelState.IsValid)
+                {
+                    db.Entry(favorite).State = EntityState.Modified;
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
+                return View(favorite);
             }
-            return View(favorite);
+            else
+            {
+                return RedirectToAction("Index", "UserOperation");
+            }
+            
         }
 
         // GET: Favorites/Delete/5
         public ActionResult Delete(int? id)
         {
-            if (id == null)
+            if (Cookies.CheckPrivilege() == true)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                Favorite favorite = db.Favorites.Find(id);
+                if (favorite == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(favorite);
             }
-            Favorite favorite = db.Favorites.Find(id);
-            if (favorite == null)
+            else
             {
-                return HttpNotFound();
+                return RedirectToAction("Index", "UserOperation");
             }
-            return View(favorite);
+            
         }
 
         // POST: Favorites/Delete/5
@@ -135,10 +192,18 @@ namespace Movie.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            Favorite favorite = db.Favorites.Find(id);
-            db.Favorites.Remove(favorite);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            if (Cookies.CheckPrivilege() == true)
+            {
+                Favorite favorite = db.Favorites.Find(id);
+                db.Favorites.Remove(favorite);
+                db.SaveChanges();
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                return RedirectToAction("Index", "UserOperation");
+            }
+            
         }
 
         protected override void Dispose(bool disposing)
