+++ Movie/Controllers/HistoriesController.cs
@@ -8,9 +8,11 @@ using System.Web;
 using System.Web.Mvc;
 using Movie.Models;
 using PagedList;
+using Movie.App_Start;
 
 namespace Movie.Controllers
 {
+    [CheckLogin]
     public class HistoriesController : Controller
     {
         private MovieContext db = new MovieContext();
@@ -18,32 +20,40 @@ namespace Movie.Controllers
         // GET: Histories
         public ActionResult Index(string UserId, string VideoId, int? page)
         {
-            var history = from m in db.Histories select m;
-            if (!String.IsNullOrEmpty(UserId))
+            if (Cookies.CheckPrivilege() == true)
             {
-                int uid = int.Parse(UserId);
-                history = history.Where(s => s.UserId == uid);
+                var history = from m in db.Histories select m;
+                if (!String.IsNullOrEmpty(UserId))
+                {
+                    int uid = int.Parse(UserId);
+                    history = history.Where(s => s.UserId == uid);
+                }
+                if (!String.IsNullOrEmpty(VideoId))
+                {
+                    int vid = int.Parse(VideoId);
+                    history = history.Where(s => s.VideoId == vid);
+                }
+                List<Total> total = new List<Total>();
+                foreach (var item in history)
+                {
+                    Total detail = new Total();
+                    detail.history = item;
+                    var user = db.Users.Where(c => c.UserId == item.UserId).FirstOrDefault();
+                    detail.user = user;
+                    var video = db.Videos.Where(c => c.VideoId == item.VideoId).FirstOrDefault();
+                    detail.video = video;
+                    total.Add(detail);
+                }
+                int pagenumber = page ?? 1;
+                int pagesize = 10;
+                IPagedList<Total> pagelist = total.ToPagedList(pagenumber, pagesize);
+                return View(pagelist);
             }
-            if (!String.IsNullOrEmpty(VideoId))
+            else
             {
-                int vid = int.Parse(VideoId);
-                history = history.Where(s => s.VideoId == vid);
+                return RedirectToAction("Index", "UserOperation");
             }
-            List<Total> total = new List<Total>();
-            foreach(var item in history)
-            {
-                Total detail = new Total();
-                detail.history = item;
-                var user = db.Users.Where(c => c.UserId == item.UserId).FirstOrDefault();
-                detail.user = user;
-                var video = db.Videos.Where(c => c.VideoId == item.VideoId).FirstOrDefault();
-                detail.video = video;
-                total.Add(detail);
-            }
-            int pagenumber = page ?? 1;
-            int pagesize = 10;
-            IPagedList<Total> pagelist = total.ToPagedList(pagenumber, pagesize);
-            return View(pagelist);
+            
         }
 
 
@@ -51,22 +61,37 @@ namespace Movie.Controllers
         // GET: Histories/Details/5
         public ActionResult Details(int? id)
         {
-            if (id == null)
+            if (Cookies.CheckPrivilege() == true)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                History history = db.Histories.Find(id);
+                if (history == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(history);
             }
-            History history = db.Histories.Find(id);
-            if (history == null)
+            else
             {
-                return HttpNotFound();
+                return RedirectToAction("Index", "UserOperation");
             }
-            return View(history);
+            
         }
 
         // GET: Histories/Create
         public ActionResult Create()
         {
-            return View();
+            if (Cookies.CheckPrivilege() == true)
+            {
+                return View();
+            }
+            else
+            {
+                return RedirectToAction("Index", "UserOperation");
+            }
         }
 
         // POST: Histories/Create
@@ -74,33 +99,49 @@ namespace Movie.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "HistoryId,UserId,VideoId,HistoryTime")] History history)
         {
-            if (ModelState.IsValid)
+            if (Cookies.CheckPrivilege() == true)
             {
-                // 如果是，则在评论表取得历史记录表中最大的历史记录编号
-                var MaxId = db.Histories.Any() ? db.Histories.Max(p => p.HistoryId) : 0;
-                // 将取得最大评论编号加一赋值给将要创建的评论
-                history.HistoryId = MaxId + 1;
-                db.Histories.Add(history);
-                db.SaveChanges();
-                return RedirectToAction("Index");
-            }
+                if (ModelState.IsValid)
+                {
+                    // 如果是，则在评论表取得历史记录表中最大的历史记录编号
+                    var MaxId = db.Histories.Any() ? db.Histories.Max(p => p.HistoryId) : 0;
+                    // 将取得最大评论编号加一赋值给将要创建的评论
+                    history.HistoryId = MaxId + 1;
+                    db.Histories.Add(history);
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
 
-            return View(history);
+                return View(history);
+            }
+            else
+            {
+                return RedirectToAction("Index", "UserOperation");
+            }
+            
         }
 
         // GET: Histories/Edit/5
         public ActionResult Edit(int? id)
         {
-            if (id == null)
+            if (Cookies.CheckPrivilege() == true)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                History history = db.Histories.Find(id);
+                if (history == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(history);
             }
-            History history = db.Histories.Find(id);
-            if (history == null)
+            else
             {
-                return HttpNotFound();
+                return RedirectToAction("Index", "UserOperation");
             }
-            return View(history);
+            
         }
 
         // POST: Histories/Edit/5
@@ -108,28 +149,44 @@ namespace Movie.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "HistoryId,UserId,VideoId,HistoryTime")] History history)
         {
-            if (ModelState.IsValid)
+            if (Cookies.CheckPrivilege() == true)
             {
-                db.Entry(history).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                if (ModelState.IsValid)
+                {
+                    db.Entry(history).State = EntityState.Modified;
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
+                return View(history);
             }
-            return View(history);
+            else
+            {
+                return RedirectToAction("Index", "UserOperation");
+            }
+            
         }
 
         // GET: Histories/Delete/5
         public ActionResult Delete(int? id)
         {
-            if (id == null)
+            if (Cookies.CheckPrivilege() == true)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                History history = db.Histories.Find(id);
+                if (history == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(history);
             }
-            History history = db.Histories.Find(id);
-            if (history == null)
+            else
             {
-                return HttpNotFound();
+                return RedirectToAction("Index", "UserOperation");
             }
-            return View(history);
+            
         }
 
         // POST: Histories/Delete/5
@@ -137,21 +194,29 @@ namespace Movie.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            // 通过Delete.cshtml页面传来的id来查询历史表
-            History history = db.Histories.Find(id);
-            // 历史表里面删除该历史记录
-            db.Histories.Remove(history);
-            var comment = db.Comments.Where(c => c.UserId == history.UserId).Where(c => c.VideoId == history.VideoId);
-            foreach (var item in comment)
+            if (Cookies.CheckPrivilege() == true)
             {
-                db.Comments.Remove(item);
+                // 通过Delete.cshtml页面传来的id来查询历史表
+                History history = db.Histories.Find(id);
+                // 历史表里面删除该历史记录
+                db.Histories.Remove(history);
+                var comment = db.Comments.Where(c => c.UserId == history.UserId).Where(c => c.VideoId == history.VideoId);
+                foreach (var item in comment)
+                {
+                    db.Comments.Remove(item);
+                }
+                var favorite = db.Favorites.Where(c => c.UserId == history.UserId).Where(c => c.VideoId == history.VideoId).FirstOrDefault();
+                db.Favorites.Remove(favorite);
+                // 数据表保存
+                db.SaveChanges();
+                // 跳转历史表首页
+                return RedirectToAction("Index");
             }
-            var favorite = db.Favorites.Where(c => c.UserId == history.UserId).Where(c => c.VideoId == history.VideoId).FirstOrDefault();
-            db.Favorites.Remove(favorite);
-            // 数据表保存
-            db.SaveChanges();
-            // 跳转历史表首页
-            return RedirectToAction("Index");
+            else
+            {
+                return RedirectToAction("Index", "UserOperation");
+            }
+            
         }
 
         protected override void Dispose(bool disposing)
