+++ Movie/Controllers/UsersController.cs
@@ -12,50 +12,73 @@ using Movie.App_Start;
 
 namespace Movie.Controllers
 {
+    [CheckLogin]
     public class UsersController : Controller
     {
         private MovieContext db = new MovieContext();
 
         // GET: Users
-        [CheckLogin]
         public ActionResult Index(string NickName, string Email,int? page)
         {
-            var users = from m in db.Users select m;
-            if (!String.IsNullOrEmpty(NickName))
+            if (Cookies.CheckPrivilege() == true)
             {
-                users = users.Where(s => s.NickName.Contains(NickName));
+                var users = from m in db.Users select m;
+                if (!String.IsNullOrEmpty(NickName))
+                {
+                    users = users.Where(s => s.NickName.Contains(NickName));
+                }
+                if (!string.IsNullOrEmpty(Email))
+                {
+                    users = users.Where(s => s.Email.Contains(Email));
+                }
+                users = users.OrderBy(c => c.UserId);
+                int pagenumber = page ?? 1;
+                int pagesize = 10;
+                IPagedList<User> pagelist = users.ToPagedList(pagenumber, pagesize);
+                return View(pagelist);
             }
-            if (!string.IsNullOrEmpty(Email))
+            else
             {
-                users = users.Where(s => s.Email.Contains(Email));
+                return RedirectToAction("Index", "UserOperation");
             }
-            users = users.OrderBy(c => c.UserId);
-            int pagenumber = page ?? 1;
-            int pagesize = 10;
-            IPagedList<User> pagelist = users.ToPagedList(pagenumber, pagesize);
-            return View(pagelist);
-       
+            
         }
 
         // GET: Users/Details/5
         public ActionResult Details(int? id)
         {
-            if (id == null)
+            if (Cookies.CheckPrivilege() == true)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                User user = db.Users.Find(id);
+                if (user == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(user);
             }
-            User user = db.Users.Find(id);
-            if (user == null)
+            else
             {
-                return HttpNotFound();
+                return RedirectToAction("Index", "UserOperation");
             }
-            return View(user);
+            
         }
 
         // GET: Users/Create
         public ActionResult Create()
         {
-            return View();
+            if (Cookies.CheckPrivilege() == true)
+            {
+                return View();
+            }
+            else
+            {
+                return RedirectToAction("Index", "UserOperation");
+            }
+            
         }
 
         // POST: Users/Create
@@ -63,33 +86,49 @@ namespace Movie.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "UserId,NickName,Password,Email,Privilege")] User user)
         {
-            if (ModelState.IsValid)
+            if (Cookies.CheckPrivilege() == true)
             {
-                // 如果是，则在用户表取得用户表中最大的用户编号
-                var MaxId = db.Users.Any() ? db.Users.Max(p => p.UserId) : 0;
-                // 将取得最大用户编号加一赋值给将要创建的用户
-                user.UserId = MaxId + 1;
-                db.Users.Add(user);
-                db.SaveChanges();
-                return RedirectToAction("Index");
-            }
+                if (ModelState.IsValid)
+                {
+                    // 如果是，则在用户表取得用户表中最大的用户编号
+                    var MaxId = db.Users.Any() ? db.Users.Max(p => p.UserId) : 0;
+                    // 将取得最大用户编号加一赋值给将要创建的用户
+                    user.UserId = MaxId + 1;
+                    db.Users.Add(user);
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
 
-            return View(user);
+                return View(user);
+            }
+            else
+            {
+                return RedirectToAction("Index", "UserOperation");
+            }
+            
         }
 
         // GET: Users/Edit/5
         public ActionResult Edit(int? id)
         {
-            if (id == null)
+            if (Cookies.CheckPrivilege() == true)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                User user = db.Users.Find(id);
+                if (user == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(user);
             }
-            User user = db.Users.Find(id);
-            if (user == null)
+            else
             {
-                return HttpNotFound();
+                return RedirectToAction("Index", "UserOperation");
             }
-            return View(user);
+            
         }
 
         // POST: Users/Edit/5
@@ -97,28 +136,44 @@ namespace Movie.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "UserId,NickName,Password,Email,Privilege")] User user)
         {
-            if (ModelState.IsValid)
+            if (Cookies.CheckPrivilege() == true)
             {
-                db.Entry(user).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                if (ModelState.IsValid)
+                {
+                    db.Entry(user).State = EntityState.Modified;
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
+                return View(user);
             }
-            return View(user);
+            else
+            {
+                return RedirectToAction("Index", "UserOperation");
+            }
+            
         }
 
         // GET: Users/Delete/5
         public ActionResult Delete(int? id)
         {
-            if (id == null)
+            if (Cookies.CheckPrivilege() == true)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                User user = db.Users.Find(id);
+                if (user == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(user);
             }
-            User user = db.Users.Find(id);
-            if (user == null)
+            else
             {
-                return HttpNotFound();
+                return RedirectToAction("Index", "UserOperation");
             }
-            return View(user);
+            
         }
 
         // POST: Users/Delete/5
@@ -126,46 +181,54 @@ namespace Movie.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            // 在用户表中查询该用户
-            User user = db.Users.Find(id);
-            var vid = db.Videos.Where(c => c.UserId == user.UserId);
-            var his = db.Histories.Where(c => c.UserId == user.UserId);
-            var com = db.Comments.Where(c => c.UserId == user.UserId);
-            var fav = db.Favorites.Where(c => c.UserId == user.UserId);
-            if(vid != null)
+            if (Cookies.CheckPrivilege() == true)
             {
-                foreach(var a in vid)
+                // 在用户表中查询该用户
+                User user = db.Users.Find(id);
+                var vid = db.Videos.Where(c => c.UserId == user.UserId);
+                var his = db.Histories.Where(c => c.UserId == user.UserId);
+                var com = db.Comments.Where(c => c.UserId == user.UserId);
+                var fav = db.Favorites.Where(c => c.UserId == user.UserId);
+                if (vid != null)
                 {
-                    db.Videos.Remove(a);
+                    foreach (var a in vid)
+                    {
+                        db.Videos.Remove(a);
+                    }
                 }
+                if (his != null)
+                {
+                    foreach (var a in his)
+                    {
+                        db.Histories.Remove(a);
+                    }
+                }
+                if (com != null)
+                {
+                    foreach (var a in com)
+                    {
+                        db.Comments.Remove(a);
+                    }
+                }
+                if (fav != null)
+                {
+                    foreach (var a in fav)
+                    {
+                        db.Favorites.Remove(a);
+                    }
+                }
+                // 删除该用户
+                db.Users.Remove(user);
+                // 数据库保存
+                db.SaveChanges();
+                // 跳转List页
+                return RedirectToAction("Index");
             }
-            if (his != null)
+            else
             {
-                foreach (var a in his)
-                {
-                    db.Histories.Remove(a);
-                }
-            }  
-            if (com != null)
-            {
-                foreach (var a in com)
-                {
-                    db.Comments.Remove(a);
-                }
-            }    
-            if (fav != null)
-            {
-                foreach (var a in fav)
-                {
-                    db.Favorites.Remove(a);
-                }
-            }  
-            // 删除该用户
-            db.Users.Remove(user);
-            // 数据库保存
-            db.SaveChanges();
-            // 跳转List页
-            return RedirectToAction("Index");
+                return RedirectToAction("Index", "UserOperation");
+            }
+            
         }
 
         protected override void Dispose(bool disposing)
